fixed contains and within to handle holes, moved tests to seperate file
